projeto/requirements.txt
selenium==4.9.1
mysql-connector-python
